Feat: remove csv output file

+++ runner/common/async_utils.py
@@ -9,10 +9,12 @@ TIMEOUT_RETCODE = 124   # Same as UNIX tiemout return code
 
 # Extend asyncio.create_subprocess_exec with timeout, so it function similarly as subprocess.run()
 async def subprocess_run(args, timeout=None, stdin=None, input=None, stdout=None, stderr=None, **kwargs):
-    proc = await asyncio.create_subprocess_exec(*args, stdin=stdin, stdout=stdout, stderr=stderr, start_new_session=True, **kwargs)
-
+    proc = None
     try:
+        proc = await asyncio.create_subprocess_exec(*args, stdin=stdin, stdout=stdout, stderr=stderr, start_new_session=True, **kwargs)
         stdout, stderr =  await asyncio.wait_for(proc.communicate(input), timeout=timeout)
+    except KeyboardInterrupt:
+        pass
     except asyncio.TimeoutError:
         # if proc.returncode is None:
         #     try:
@@ -27,7 +29,7 @@ async def subprocess_run(args, timeout=None, stdin=None, input=None, stdout=None
         #         pass
         return ("", f'Timeout after {timeout}s'.encode(), TIMEOUT_RETCODE)
     finally:
-        if proc.returncode is None:
+        if proc is not None and proc.returncode is None:
             try:
                 os.killpg(os.getpgid(proc.pid), signal.SIGTERM)
             except:
